Clang:  Fix warnings when building with clang 10

+++ examples/GUI/WindowsDemo.h
@@ -116,7 +116,7 @@ public:
 
         Random random;
 
-        auto size = 10.0f + random.nextInt (30);
+        auto size = 10.0f + (float) random.nextInt (30);
 
         ballBounds.setBounds (random.nextFloat() * 100.0f,
                               random.nextFloat() * 100.0f,
@@ -142,10 +142,10 @@ public:
     {
         ballBounds += direction;
 
-        if (ballBounds.getX() < 0)                      direction.x =  std::abs (direction.x);
-        if (ballBounds.getY() < 0)                      direction.y =  std::abs (direction.y);
-        if (ballBounds.getRight()  > getParentWidth())  direction.x = -std::abs (direction.x);
-        if (ballBounds.getBottom() > getParentHeight()) direction.y = -std::abs (direction.y);
+        if (ballBounds.getX() < 0)                              direction.x =  std::abs (direction.x);
+        if (ballBounds.getY() < 0)                              direction.y =  std::abs (direction.y);
+        if (ballBounds.getRight()  > (float) getParentWidth())  direction.x = -std::abs (direction.x);
+        if (ballBounds.getBottom() > (float) getParentHeight()) direction.y = -std::abs (direction.y);
 
         setBounds (ballBounds.getSmallestIntegerContainer());
     }
