+++ extras/Projucer/Source/Utility/UI/PropertyComponents/jucer_ColourPropertyComponent.h
@@ -71,7 +71,7 @@ private:
                                 Colour (0xffffffff).overlaidWith (colour));
 
             g.setColour (Colours::white.overlaidWith (colour).contrasting());
-            g.setFont (Font (getHeight() * 0.6f, Font::bold));
+            g.setFont (Font ((float) getHeight() * 0.6f, Font::bold));
             g.drawFittedText (colour.toDisplayString (true), getLocalBounds().reduced (2, 1),
                               Justification::centred, 1);
         }
