+++ modules/juce_gui_basics/layout/juce_ComponentAnimator.cpp
@@ -175,8 +175,8 @@ public:
         void paint (Graphics& g) override
         {
             g.setOpacity (1.0f);
-            g.drawImageTransformed (image, AffineTransform::scale (getWidth()  / (float) jmax (1, image.getWidth()),
-                                                                   getHeight() / (float) jmax (1, image.getHeight())), false);
+            g.drawImageTransformed (image, AffineTransform::scale ((float) getWidth()  / (float) jmax (1, image.getWidth()),
+                                                                   (float) getHeight() / (float) jmax (1, image.getHeight())), false);
         }
 
     private:
